refactor: 로그인, 토큰 함수 폴더이동
