+++ client/src/components/@layout/header/MemberMenu.tsx
@@ -1,7 +1,7 @@
 import { styled } from 'styled-components';
 import LoginSignup from '@/components/@layout/header/LoginSignup';
 import UserProfile from '@/components/@layout/header/UserProfile';
-import useIsLoggedIn from '@/hooks/useIsLoggedIn';
+import useIsLoggedIn from '@/utils/isLoggedIn';
 
 function MemberMenu({ position }: { position: number }) {
   const isLoggedIn = useIsLoggedIn();
