+++ client/src/components/comments/CommentForm.tsx
@@ -4,7 +4,7 @@ import { BiPaperPlane } from 'react-icons/bi';
 import { useParams } from 'react-router-dom';
 import { styled } from 'styled-components';
 import { PostComment } from '@/api/api';
-import useIsLoggedIn from '@/hooks/useIsLoggedIn';
+import useIsLoggedIn from '@/utils/isLoggedIn';
 
 function CommentForm() {
   const [content, setContent] = useState('');
