+++ client/src/components/mediaDetail/bookmark/Bookmark.tsx
@@ -4,8 +4,8 @@ import { BsHeart, BsHeartFill } from 'react-icons/bs';
 import { useNavigate } from 'react-router-dom';
 import { GetIsBookmark, PostBookmark } from '@/api/api';
 import BookmarkLoading from '@/components/mediaDetail/bookmark/BookmarkLoading';
-import useIsLoggedIn from '@/hooks/useIsLoggedIn';
 import { S_IconWrapper } from '@/styles/style';
+import useIsLoggedIn from '@/utils/isLoggedIn';
 import { notifyError, notifyWithIcon } from '@/utils/notify';
 
 function Bookmark({ contentId }: { contentId: string }) {
