+++ client/src/pages/Main.tsx
@@ -2,8 +2,8 @@ import { useSearchParams } from 'react-router-dom';
 import MainSliderSection from '@/components/@common/slider/MainSliderSection';
 import MainBanner from '@/components/@layout/banners/MainBanner';
 import { REFRSH_TOKEN_DURATION } from '@/constant/constantValue.ts';
-import useIsLoggedIn from '@/hooks/useIsLoggedIn';
-import { useTokens } from '@/hooks/useTokens.ts';
+import { areTokens } from '@/utils/areTokens';
+import useIsLoggedIn from '@/utils/isLoggedIn';
 import { scrollToTop } from '@/utils/scrollToTop.ts';
 
 function Main() {
@@ -12,7 +12,7 @@ function Main() {
   const accessToken = searchParams.get('access_token');
   const refreshToken = searchParams.get('refresh_token');
   if (!isLoggedIn && accessToken) {
-    useTokens(accessToken, refreshToken);
+    areTokens(accessToken, refreshToken);
     const expiration = new Date();
     expiration.setMinutes(expiration.getMinutes() + REFRSH_TOKEN_DURATION);
     localStorage.setItem('expiration', expiration.toISOString());
